Improve cloud installation process
* Save results of `run-local.sh` execution into a dedicated log (by
  default `arp-vm`
* Check for `google-ads.yaml` file both in solution folder
  (`app-reporting-pack`) and in $HOME directory
* Change `config_file` variable to `app/app_reporting_pack.yaml` in
  `run-local.sh` - it was causing errors during config generation and
  resulted in skipping certain sections (i.e. iOS SKAN)

+++ gcp/setup.sh
@@ -49,7 +49,13 @@ deploy_files() {
   # NOTE: if an error "module 'sys' has no attribute 'maxint'" occures, run this: `pip3 install -U crcmod`
   gsutil -m rsync -r -x ".*/__pycache__/.*|\..*" ./../app $GCS_BASE_PATH
   gsutil -h "Content-Type:text/plain" cp ./../app/*.yaml $GCS_BASE_PATH/
-  gsutil -h "Content-Type:text/plain" cp ./../google-ads.yaml $GCS_BASE_PATH/google-ads.yaml
+  if [[ -f ./../google-ads.yaml ]]; then
+    gsutil -h "Content-Type:text/plain" cp ./../google-ads.yaml $GCS_BASE_PATH/google-ads.yaml
+  elif [[ -f $HOME/google-ads.yaml ]]; then
+    gsutil -h "Content-Type:text/plain" cp $HOME/google-ads.yaml $GCS_BASE_PATH/google-ads.yaml
+  else
+    echo "Please upload google-ads.yaml"
+  fi
 }
